borg: don't hardcode ssh port default (DEBBUG-993951)
Adding the port parameter to the borg handler with a default value of 22
is causing issues for users who had been using the standard ssh config
file to set the remote port, since the -p cli argument overrides the
configuration. This removes the default port number and only uses it in
the connection string if defined.